+++ src/main/java/org/swisspush/redisques/util/QueueStatisticsCollector.java
@@ -29,6 +29,7 @@ import java.util.Map.Entry;
 import java.util.Set;
 import java.util.concurrent.ConcurrentHashMap;
 import java.util.concurrent.Semaphore;
+import java.util.concurrent.atomic.AtomicBoolean;
 import java.util.concurrent.atomic.AtomicLong;
 import java.util.function.BiConsumer;
 
@@ -65,6 +66,7 @@ public class QueueStatisticsCollector {
     private static final Logger log = LoggerFactory.getLogger(QueueStatisticsCollector.class);
 
     private static final String STATSKEY = "redisques:stats";
+    public static final AtomicBoolean CODECS_REGISTERED = new AtomicBoolean(false);
     private final static String QUEUE_FAILURES = "failures";
     private final static String QUEUE_BACKPRESSURE = "backpressureTime";
     private final static String QUEUE_SLOWDOWNTIME = "slowdownTime";
@@ -86,6 +88,7 @@ public class QueueStatisticsCollector {
     private MessageConsumer<QueueSizeInfoMap> queueSizeConsumer;
     private Long speedStatisticsTimerId;
 
+
     public QueueStatisticsCollector(
             RedisService redisService,
             KeyspaceHelper keyspaceHelper,
@@ -103,7 +106,7 @@ public class QueueStatisticsCollector {
         this.configurationProvider = configurationProvider;
         this.keyspaceHelper = keyspaceHelper;
 
-        vertx.eventBus().registerDefaultCodec(QueueSizeInfoMap.class, new QueueSizeInfoMapCodec());
+        registerCodecs(vertx);
         speedStatisticsScheduler(speedIntervalSec);
         this.queueSizeConsumer = vertx.eventBus().consumer(keyspaceHelper.getQueueStatisticQueueSizeSyncKey(),
                 (Handler<Message<QueueSizeInfoMap>>) event ->
@@ -113,6 +116,12 @@ public class QueueStatisticsCollector {
                                 approximateQueueSize.put(key, value);
                             }
                         }));
+    }
+
+    public static void registerCodecs(Vertx vertx) {
+        if (CODECS_REGISTERED.compareAndSet(false, true)) {
+            vertx.eventBus().registerDefaultCodec(QueueSizeInfoMap.class, new QueueSizeInfoMapCodec());
+        }
     }
 
     /**