+++ src/test/java/org/swisspush/redisques/migration/MigrateToolTest.java
@@ -21,6 +21,7 @@ import org.swisspush.redisques.migration.tasks.Task;
 import org.swisspush.redisques.util.DefaultRedisquesConfigurationProvider;
 import org.swisspush.redisques.util.QueueConfiguration;
 import org.swisspush.redisques.util.QueueConfigurationProvider;
+import org.swisspush.redisques.util.QueueStatisticsCollector;
 import org.swisspush.redisques.util.RedisquesConfiguration;
 import org.swisspush.redisques.util.TestMemoryUsageProvider;
 import redis.clients.jedis.Jedis;
@@ -44,6 +45,7 @@ public class MigrateToolTest extends AbstractTestCase {
     @Before
     public void deployRedisques(TestContext context) {
         vertx = Vertx.vertx();
+        QueueStatisticsCollector.CODECS_REGISTERED.set(false);
         Async async = context.async();
         QueueConfigurationProvider.reset();
         RedisquesConfiguration rqConfig =  RedisquesConfiguration.with()